requirements.txt
aiosmtplib==2.0.1
alembic==1.9.1
anyio==3.6.2
attrs==22.2.0
bcrypt
cachetools==4.2.4
email-validator
fastapi
fastapi-mail
Jinja2==3.1.2
passlib==1.7.4
psycopg==3.1.9
psycopg-binary
pydantic==1.10.4
pytest==7.2.0
pytest-postgresql
python-dateutil
python-decouple
python-dotenv
python-jose
SQLAlchemy
starlette
tenacity
uvicorn
cloudinary
pytest-asyncio
pytest-mock
python-slugify